validación de parámetros

+++ src/controllers/categories-controller.js
@@ -6,7 +6,7 @@ const getCategories = async (req, res) => {
     if(response.rows.length > 0){
         res.status(200).json(response.rows);
     }else{
-        res.status(404).json({error: 'not found'});
+        res.status(404).json({error: 'No se encontró'});
     }
 };
 
@@ -17,10 +17,10 @@ const getCategorieById = async (req, res) => {
         if(response.rows.length > 0){
             res.status(200).json(response.rows[0]);
         }else{
-            res.status(404).json({error: 'not found'});
+            res.status(404).json({error: 'No se encontró'});
         }
     }else{
-        res.status(400).json({error: 'invalid parameter'});
+        res.status(400).json({error: 'Parámetro inválido'});
     }
 };
 
@@ -39,13 +39,19 @@ const createCategory = async(req, res) => {
 
 const deleteCategory = async(req, res) => {
     const {id_category} = req.body
-    await database.query('DELETE FROM categories WHERE id = $1',[id_category],function(err, result, fields) {
-        if (err) {
-            res.status(400).json({error: 'Algo salió mal'});
-        }else{
-            res.status(200).json({message: 'Categoría eliminada satisfactoriamente'});
-        }
-    });
+    const check_category = await database.query('SELECT * FROM categories WHERE id = $1')
+    if (check_category.rowCount > 0){
+        await database.query('DELETE FROM categories WHERE id = $1',[id_category],function(err, result, fields) {
+            if (err) {
+                res.status(400).json({error: 'Algo salió mal'});
+            }else{
+                res.status(200).json({message: 'Categoría eliminada satisfactoriamente'});
+            }
+        });
+    } else{
+        res.status(404).json({error: 'No se encontró la categoría'});
+    }
+
 }
 
 module.exports = {
